+++ src/app/(studios)/color-studio-v2/layout.tsx
@@ -1,5 +1,4 @@
 import type { Metadata } from 'next'
-import { StudioHeader } from './_shared/components/StudioHeader'
 import { ColorStudioProvider } from './_shared/contexts/ColorStudioContext'
 
 export const metadata: Metadata = {
@@ -11,8 +10,8 @@ export const metadata: Metadata = {
 export default function ColorStudioV2Layout({ children }: { children: React.ReactNode }) {
   return (
     <ColorStudioProvider>
-      <div className='my-10 h-screen overflow-auto'>
-        <StudioHeader />
+      <div className='h-full overflow-auto'>
+        {/* <StudioHeader /> */}
         {children}
       </div>
     </ColorStudioProvider>